requirements.txt
google-genai
python-dateutil
requests